requirements.txt
opencv-python
scikit-learn
tqdm
pyyaml
argparse
json